Create module for naming things
